Migrate health check routes to Effect patterns
- Convert all health check endpoints to use effectToExpress adapter
- Replace try/catch with Effect's type-safe error handling
- Extract helper functions using Effect generators for composability
- Maintain all existing functionality with improved error safety
- Add proper error channels for database connectivity checks
