+++ administration/server/src/index.ts
@@ -2,11 +2,8 @@ import dotenv from 'dotenv';
 
 import cors from 'cors';
 import express from 'express';
-import * as fs from 'fs';
 import helmet from 'helmet';
 import morgan from 'morgan';
-import * as os from 'os';
-import * as path from 'path';
 import { auditMiddleware as globalAuditMiddleware } from './middleware/auditLoggingMiddleware';
 import { errorHandler } from './middleware/errorHandler';
 import { apiLimiter } from './middleware/rateLimiting';
@@ -15,11 +12,12 @@ import { auditRoutes } from './routes/auditRoutes';
 import { discordRoutes } from './routes/discordRoutes';
 import { eventbriteRoutes } from './routes/eventbriteRoutes';
 import { eventRoutes } from './routes/eventRoutes';
+import { healthCheckRouter } from './routes/healthCheck';
 import { klaviyoRoutes } from './routes/klaviyoRoutes';
 import { memberRoutes } from './routes/memberRoutes';
 import { patreonRoutes } from './routes/patreonRoutes';
 import { webhookRoutes } from './routes/webhookRoutes';
-import { getDatabaseInfo, getDb, initialize } from './services/DatabaseService';
+import { initialize } from './services/DatabaseService';
 import { jobScheduler } from './services/JobScheduler';
 import logger from './utils/logger';
 
@@ -66,6 +64,7 @@ if (process.env.NODE_ENV === 'production') {
 }
 
 // Routes
+app.use('/health', healthCheckRouter);
 app.use('/api/members', memberRoutes);
 app.use('/api/events', eventRoutes);
 app.use('/api/webhooks', webhookRoutes);
@@ -75,221 +74,6 @@ app.use('/api/patreon', patreonRoutes);
 app.use('/api/discord', discordRoutes);
 app.use('/api/applications', applicationRoutes);
 app.use('/api/audit', auditRoutes);
-
-
-if (process.env.NODE_ENV !== 'production') {
-// Health check with optional debug info
-  app.get('/health', async (req, res) => {
-    try {
-      // Check database connectivity
-      const db = getDb();
-      await db.selectFrom('payment_statuses').select('id').limit(1).execute();
-
-      const basicHealth = {
-        status: 'healthy',
-        timestamp: new Date().toISOString(),
-        version: process.env.npm_package_version || '1.0.0',
-        database: 'connected',
-        environment: process.env.NODE_ENV || 'development',
-      };
-
-      // Add comprehensive debug information
-      const dbPath = process.env.DATABASE_PATH?.replace('file:', '');
-      if (!dbPath) {
-        throw new Error('DATABASE_PATH is not set');
-      }
-      const absoluteDbPath = path.resolve(dbPath);
-
-      let dbStats: {
-        size: number;
-        sizeFormatted: string;
-        modified: string;
-        created: string;
-      } | null = null;
-      let dbSize = 'unknown';
-      try {
-        const stats = fs.statSync(absoluteDbPath);
-        dbStats = {
-          size: stats.size,
-          sizeFormatted: `${(stats.size / 1024 / 1024).toFixed(2)} MB`,
-          modified: stats.mtime.toISOString(),
-          created: stats.birthtime.toISOString(),
-        };
-        dbSize = dbStats.sizeFormatted;
-      } catch (error) {
-        // Database file might not exist or be accessible
-      }
-
-      // Get database table counts
-      let tableCounts: any = {};
-      try {
-        const memberResult = await db
-          .selectFrom('members')
-          .select(db.fn.count('id').as('count'))
-          .executeTakeFirst();
-        const eventResult = await db
-          .selectFrom('events')
-          .select(db.fn.count('id').as('count'))
-          .executeTakeFirst();
-        // Note: 'memberships' table exists in our schema
-        const membershipResult = await db
-          .selectFrom('memberships' as any)
-          .select(db.fn.count('id').as('count'))
-          .executeTakeFirst();
-
-        tableCounts = {
-          members: memberResult?.count || 0,
-          events: eventResult?.count || 0,
-          memberships: membershipResult?.count || 0,
-        };
-      } catch (error) {
-        tableCounts = { error: 'Could not retrieve table counts' };
-      }
-
-      // Get git information
-      let gitInfo = {};
-      try {
-        const gitHeadPath = path.resolve('.git/HEAD');
-        if (fs.existsSync(gitHeadPath)) {
-          const head = fs.readFileSync(gitHeadPath, 'utf8').trim();
-          if (head.startsWith('ref: ')) {
-            const refPath = head.substring(5);
-            const commitPath = path.resolve('.git', refPath);
-            if (fs.existsSync(commitPath)) {
-              const commit = fs.readFileSync(commitPath, 'utf8').trim();
-              gitInfo = {
-                branch: refPath.replace('refs/heads/', ''),
-                commit: commit.substring(0, 7),
-                fullCommit: commit,
-              };
-            }
-          }
-        }
-      } catch (error) {
-        gitInfo = { error: 'Could not retrieve git information' };
-      }
-
-      const debugInfo = {
-        ...basicHealth,
-        debug: {
-          application: {
-            name: 'Club Management System',
-            version: '1.0.0',
-            environment: process.env.NODE_ENV || 'development',
-            port: process.env.PORT || 3000,
-            uptime: `${Math.floor(process.uptime())} seconds`,
-          },
-          database: {
-            type: 'SQLite',
-            path: absoluteDbPath,
-            exists: fs.existsSync(absoluteDbPath),
-            stats: dbStats,
-            size: dbSize,
-            tableCounts,
-          },
-          system: {
-            platform: os.platform(),
-            arch: os.arch(),
-            hostname: os.hostname(),
-            release: os.release(),
-            totalMemory: `${(os.totalmem() / 1024 / 1024 / 1024).toFixed(2)} GB`,
-            freeMemory: `${(os.freemem() / 1024 / 1024 / 1024).toFixed(2)} GB`,
-            cpus: os.cpus().length,
-            loadAverage: os.loadavg(),
-          },
-          process: {
-            pid: process.pid,
-            version: process.version,
-            nodeVersion: process.versions.node,
-            platform: process.platform,
-            arch: process.arch,
-            memoryUsage: process.memoryUsage(),
-            uptime: process.uptime(),
-            cwd: process.cwd(),
-          },
-          environment: {
-            NODE_ENV: process.env.NODE_ENV,
-            DATABASE_PATH: process.env.DATABASE_PATH ? '[CONFIGURED]' : '[NOT SET]',
-            KLAVIYO_API_KEY: process.env.KLAVIYO_API_KEY ? '[CONFIGURED]' : '[NOT SET]',
-            DISCORD_BOT_TOKEN: process.env.DISCORD_BOT_TOKEN ? '[CONFIGURED]' : '[NOT SET]',
-            PATREON_CLIENT_ID: process.env.PATREON_CLIENT_ID ? '[CONFIGURED]' : '[NOT SET]',
-            PORT: process.env.PORT || '[DEFAULT: 3000]',
-          },
-          git: gitInfo,
-          network: Object.entries(os.networkInterfaces())
-            .filter(([name]) => !name.startsWith('lo')) // Filter out loopback
-            .map(([name, addresses]) => ({
-              name,
-              addresses:
-                addresses?.filter((addr) => addr.family === 'IPv4').map((addr) => addr.address) ||
-                [],
-            })),
-          time: {
-            iso: new Date().toISOString(),
-            unix: Math.floor(Date.now() / 1000),
-            local: new Date().toString(),
-            utc: new Date().toUTCString(),
-          },
-        },
-      };
-
-      return res.json(debugInfo);
-    } catch (error) {
-      return res.status(503).json({
-        status: 'unhealthy',
-        timestamp: new Date().toISOString(),
-        error: 'Database connection failed',
-        database: 'disconnected',
-      });
-    }
-  });
-}
-
-// Environment check endpoint
-app.get('/health/env', (req, res) => {
-  const actualValues = ['DATABASE_PATH'];
-  const requiredVars = ['JWT_SECRET'];
-  const optionalVars = [
-    'KLAVIYO_API_KEY',
-    'DISCORD_BOT_TOKEN',
-    'EVENTBRITE_API_TOKEN',
-    'PATREON_CLIENT_ID',
-  ];
-
-  const env = {
-    values: {} as Record<string, any>,
-    required: {} as Record<string, boolean>,
-    optional: {} as Record<string, boolean>,
-    issues: [] as string[],
-  };
-
-  actualValues.forEach((varName) => {
-    const exists = process.env[varName];
-    env.values[varName] = exists;
-    if (!exists) {
-      env.issues.push(`Missing required environment variable: ${varName}`);
-    }
-  });
-  requiredVars.forEach((varName) => {
-    const exists = !!process.env[varName];
-    env.required[varName] = exists;
-    if (!exists) {
-      env.issues.push(`Missing required environment variable: ${varName}`);
-    }
-  });
-
-  optionalVars.forEach((varName) => {
-    env.optional[varName] = !!process.env[varName];
-  });
-
-  const hasIssues = env.issues.length > 0;
-  res.status(hasIssues ? 422 : 200).json({
-    status: hasIssues ? 'configuration_issues' : 'ok',
-    timestamp: new Date().toISOString(),
-    environment: env,
-    database: getDatabaseInfo(),
-  });
-});
 
 // Queue status endpoint
 app.get('/api/queue/status', async (req, res) => {
