+++ administration/server/src/services/DatabaseService.ts
@@ -24,7 +24,6 @@ if (!dbPath) {
 function initializeDatabase(): void {
   if (_initialized) return;
 
-
   // Ensure data directory exists
   const dbDir = path.dirname(dbPath);
   if (!fs.existsSync(dbDir)) {
@@ -78,52 +77,4 @@ export function prepare(sql: string): Database.Statement {
     initializeDatabase();
   }
   return _sqliteDb.prepare(sql);
-}
-
-export function getDatabaseInfo() {
-  if (!_initialized) {
-    initializeDatabase();
-  }
-
-  const absoluteDbPath = path.resolve(dbPath);
-
-  // Get SQLite version information
-  const sqliteVersion = _sqliteDb.pragma('sqlite_version', { simple: true });
-  const userVersion = _sqliteDb.pragma('user_version', { simple: true });
-  const schemaVersion = _sqliteDb.pragma('schema_version', { simple: true });
-
-  // Get file system information
-  let fileInfo: any = null;
-  try {
-    const stats = fs.statSync(absoluteDbPath);
-    fileInfo = {
-      size: stats.size,
-      sizeFormatted: `${(stats.size / 1024 / 1024).toFixed(2)} MB`,
-      modified: stats.mtime.toISOString(),
-      created: stats.birthtime.toISOString(),
-      path: absoluteDbPath,
-    };
-  } catch (error) {
-    fileInfo = { error: 'Could not retrieve database file info' };
-  }
-
-  // Get database statistics
-  const pageCount = _sqliteDb.pragma('page_count', { simple: true }) as number;
-  const pageSize = _sqliteDb.pragma('page_size', { simple: true }) as number;
-  const cacheSize = _sqliteDb.pragma('cache_size', { simple: true }) as number;
-
-  return {
-    sqlite: {
-      version: sqliteVersion,
-      userVersion: userVersion,
-      schemaVersion: schemaVersion,
-    },
-    file: fileInfo,
-    statistics: {
-      pageCount: pageCount,
-      pageSize: pageSize,
-      cacheSize: cacheSize,
-      totalPages: pageCount * pageSize,
-    },
-  };
 }
